requirements.txt
playwright==1.47.0
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
